build: Fix patches for Linguist (#12)
When generating patches for Linguist, one needs to use the `--no-prefix`
option for `git diff`, otherwise it interprets the `a/` and `b/` prefixes as
part of the directory structure.

This avoids the need to point to a separate fork of linguist for the
`.bazelignore` change. I'm also fixing another issue that we're running
into when attempting to consistently use Language names.